feat: implement `sync_cfg` and `async_cfg`
When the `#[async_generic]` macro has a `sync_cfg(...)` argument,
the parameters to that argument are passed into a `#[cfg(...)]` block
for the sync version of the function.

When the `#[async_generic]` macro has an `async_cfg(...)` argument,
the parameters to that argument are passed into a `#[cfg(...)]` block
for the async version of that function.

Together, this gives users complete control over which versions
of the functions are compiled, when.